+++ src/config/auth.ts
@@ -1,4 +1,4 @@
 export default {
   secretToken: process.env.AUTH_SECRET || "default",
-  expiresInToken: Number(process.env.AUTH_EXPIRES_TOKEN) || 360,
+  expiresInToken: Number(process.env.AUTH_EXPIRES_TOKEN) || 3600,
 };