+++ src/domain/ports/IShortUrlsRepository.ts
@@ -4,5 +4,6 @@ import type { ICreateShortUrlDTO } from "./ICreateShortUrlDTO";
 export interface IShortUrlsRepository {
   create(data: ICreateShortUrlDTO): Promise<IShortUrl>;
   findById(id: string): Promise<IShortUrl | null>;
-  updateClickCount(id: string): Promise<null>;
+  listByUserId(userId: string): Promise<IShortUrl[]>;
+  updateClickCount(id: string): Promise<void>;
 }